Refactor code to use Redux toolkit and upgrade React to concurrent mode
The changes include a major refactoring of the actions, reducers, and parts of the components to use Redux toolkit. Redux is refactored to use @reduxjs/toolkit and some ThunkActions have been replaced with createAsyncThunk. React is also upgraded to concurrent mode (React 18). Additionally, framer-motion library was added to provide animations for the components.